Added padding to cached items, so that we can cache items that draw outside its bounds
It is currently hard coded to 16px on each side, meaning that we will
get glitches for items that draw more than 16px outside its bounds.
Eventually, we will calculate the padding for each individual primitive,
but since the user will already decide which items should be cached,
it is easiest to start with them also stating this value manually first.